Added a printHeader function to trainmonitor that will print formatted messages example:  printHeader("Done loading sections"); will print  -------------------Done loading sections-------------------
As for the crossovers: I have added it to the sections.txt and also I have rearraged the sections.txt in numerical order.
I saw that there are 2 entries of section 3-3 with one being an endpoint and the other being a straight section... I we need to resolve how to capture 3-3 as one record, so what is this supposed to really be? is the 4-9 record that links 3-3 enough? if so we can keep the 3-3 record that shows it as an endpoint... This is what I did for now because along with 4-9,  1-5 & 3-2 also keep track of 3-3.